// Airbnb-front/src/app/components/models/onlineExperience.ts
import { Time } from "@angular/common";

export class onlineExperience{
    onlineExperienceID: number;
    onlineExperienceDescription: string;
    onlineExperienceType: any;
    onlineExperienceLanguage: string;
    onlineExperienceLink: string;
    onlineExperienceDate: Date;
    startTime: Date;
    endTime: Date;
    hostName: string;
    hostEmail: string;
    onlineExperienceStatus: any;
    onlineExperienceImage: any;
}